feat: move versionmismatch filtering into API
[ORQSDK-843]
The previous location filtered warnigns only when using the CLI.
